requirements.txt
Django==4.2.2

# Utils:
# postgres driver:
psycopg2==2.9.6
# django settings organization:
python-decouple==3.8
dj-database-url==2.0.0
# code coverage:
coverage==7.2.7
# logging @ sentry:
raven==6.10.0

# Apps:
# code testing:
django-nose==1.4.7
# rich-text editor:
django-ckeditor==6.5.1
django-js-asset==2.0.0
# form presentation:
django-widget-tweaks==1.4.12
